Rename collectErrors to collectParseErrors
