--- src/shared/collectErrors/index.ts
@@ -1,1 +1,0 @@
-export { collectErrors } from './CollectErrors'
