+++ src/shared/parseError/ParseError.ts
@@ -1,8 +1,18 @@
-import { ParseError } from '@shared/ParseError'
+export class ParseError extends Error {
+  private constructor(error: Error) {
+    super(error.message)
+
+    Object.setPrototypeOf(this, ParseError.prototype)
+  }
+
+  public static fromError(error: Error) {
+    return new this(error)
+  }
+}
 
 type Validator<T> = (val: T) => ParseError | null
 
-export const collectErrors = <T>(
+export const collectParseErrors = <T>(
   validators: Validator<T>[],
   val: T
 ): ParseError[] =>
