+++ src/shared/parseError/index.ts
@@ -1,0 +1,1 @@
+export * from './ParseError'
